use global rank
