+++ pytorch_lightning/loggers/base.py
@@ -8,7 +8,7 @@ from typing import Union, Optional, Dict, Iterable, Any, Callable, List, Sequenc
 import numpy as np
 import torch
 
-from pytorch_lightning.utilities.distributed import rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
 
 
 class LightningLoggerBase(ABC):
@@ -238,16 +238,6 @@ class LightningLoggerBase(ABC):
         self.save()
 
     @property
-    def rank(self) -> int:
-        """Process rank. In general, metrics should only be logged by the process with rank 0."""
-        return self._rank
-
-    @rank.setter
-    def rank(self, value: int) -> None:
-        """Set the process rank."""
-        self._rank = value
-
-    @property
     @abstractmethod
     def name(self) -> str:
         """Return the experiment name."""
@@ -297,11 +287,6 @@ class LoggerCollection(LightningLoggerBase):
     @rank_zero_only
     def close(self) -> None:
         [logger.close() for logger in self._logger_iterable]
-
-    @LightningLoggerBase.rank.setter
-    def rank(self, value: int) -> None:
-        for logger in self._logger_iterable:
-            logger.rank = value
 
     @property
     def name(self) -> str:
