+++ pytorch_lightning/loggers/comet.py
@@ -26,7 +26,7 @@ from torch import is_tensor
 from pytorch_lightning import _logger as log
 from pytorch_lightning.loggers.base import LightningLoggerBase
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.distributed import rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
 
 
 class CometLogger(LightningLoggerBase):
