+++ pytorch_lightning/loggers/tensorboard.py
@@ -15,7 +15,7 @@ from torch.utils.tensorboard import SummaryWriter
 
 from pytorch_lightning import _logger as log
 from pytorch_lightning.loggers.base import LightningLoggerBase
-from pytorch_lightning.utilities.distributed import rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
 
 
 class TensorBoardLogger(LightningLoggerBase):
