+++ pytorch_lightning/loggers/trains.py
@@ -20,7 +20,7 @@ except ImportError:  # pragma: no-cover
 
 from pytorch_lightning import _logger as log
 from pytorch_lightning.loggers.base import LightningLoggerBase
-from pytorch_lightning.utilities.distributed import rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
 
 
 class TrainsLogger(LightningLoggerBase):
