+++ pytorch_lightning/loggers/wandb.py
@@ -16,7 +16,7 @@ except ImportError:  # pragma: no-cover
                       ' install it with `pip install wandb`.')
 
 from pytorch_lightning.loggers.base import LightningLoggerBase
-from pytorch_lightning.utilities.distributed import rank_zero_only
+from pytorch_lightning.utilities import rank_zero_only
 
 
 class WandbLogger(LightningLoggerBase):
