+++ pytorch_lightning/trainer/distrib_data_parallel.py
@@ -123,7 +123,7 @@ from pytorch_lightning import _logger as log
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loggers import LightningLoggerBase
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.warnings import set_proc_rank, rank_zero_warn
+from pytorch_lightning.utilities.distributed import set_proc_rank, rank_zero_warn
 
 try:
     from apex import amp
@@ -324,14 +324,9 @@ class TrainerDDPMixin(ABC):
         elif self.use_ddp2:
             self.proc_rank = self.node_rank
             self.world_size = self.num_nodes
+
         # set warning rank
         set_proc_rank(self.proc_rank)
-
-        # let the exp know the rank to avoid overwriting logs
-        if isinstance(self.logger, LightningLoggerBase):
-            self.logger.rank = self.proc_rank
-        if isinstance(self.checkpoint_callback, ModelCheckpoint):
-            self.checkpoint_callback.rank = self.proc_rank
 
         # set up server using proc 0's ip address
         # try to init for 20 times at max in case ports are taken
