+++ pytorch_lightning/trainer/logging.py
@@ -33,7 +33,6 @@ class TrainerLoggingMixin(ABC):
                 version=self.slurm_job_id,
                 name='lightning_logs'
             )
-            self.logger.rank = 0
         elif logger is False:
             self.logger = None
         else:
@@ -41,7 +40,6 @@ class TrainerLoggingMixin(ABC):
                 self.logger = LoggerCollection(logger)
             else:
                 self.logger = logger
-            self.logger.rank = 0
 
     def log_metrics(self, metrics, grad_norm_dic, step=None):
         """Logs the metric dict passed in.
