+++ pytorch_lightning/utilities/__init__.py
@@ -1,3 +1,3 @@
 """General utilities"""
 
-from pytorch_lightning.utilities.warnings import rank_zero_warn
+from pytorch_lightning.utilities.distributed import rank_zero_only, rank_zero_warn
